requirements-dev.txt
# 测试框架
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-xdist==3.3.1
requests-mock==1.11.0

# 静态检查
flake8==6.0.0
mypy==1.5.1
black==23.7.0
isort==5.12.0

# 文档生成
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
mkdocs==1.5.2
mkdocs-material==9.1.21

# 构建工具
wheel==0.41.2
setuptools==68.0.0
twine==4.0.2

# 开发辅助
ipython==8.14.0
pre-commit==3.3.3
python-dotenv==1.0.0

# 音频处理额外依赖
numpy==1.25.2
sounddevice==0.4.6
librosa==0.10.1

# UI开发依赖
pygame==2.5.0
PyOpenGL==3.1.7

# 可选GPU加速
cupy-cuda11x==12.2.0 ; sys_platform != 'win32'
nvidia-ml-py3==7.352.0